Bug #179273 fix compiler warning message for sc8830
	[bug number ] 179273
	[root cause ] no
	[changes ]
	[side effects ] no
	[reviewers ] steve.zhan

+++ arch/arm/mach-sc/efuse.c
@@ -525,7 +525,7 @@ static int fuse_debug_get(void *data, u64 * val)
 	struct sci_fuse *p = data;
 
 	if (p->is_adie_fuse == 2) {
-		*val = read_write_i2s_switch_pin(1, val, p);
+		*val = read_write_i2s_switch_pin(1, (int)val, p);
 		return 0;
 	}
